skills/image_skill/requirements.txt
requests==2.27.1
Pillow==9.1.0
fastapi==0.73.0
uvicorn==0.17.4
